+++ libs/backend/application/action/src/action.application.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common'
 import { ActionApplicationController } from './action.application.controller'
 import { ActionApplicationService } from './action.application.service'
 import { ActionDomainModule } from '@codelab/backend/domain/action'
+import { CodelabLoggerModule } from '@codelab/backend/infra/adapter/logger'
 
 @Module({
-  imports: [ActionDomainModule],
+  imports: [ActionDomainModule, CodelabLoggerModule],
   controllers: [ActionApplicationController],
   providers: [ActionApplicationService],
 })
