+++ libs/frontend/test/cypress/nextjs-auth0/src/index.ts
@@ -1,4 +1,4 @@
 import './register.cypress'
 
 export * from './commands/login-and-setup-e2e-data'
-export * from './nextjs-auth0.commands'
+export * from './commands/nextjs-auth0.commands'
