+++ libs/frontend/test/cypress/utils/src/index.ts
@@ -1,3 +1,4 @@
+import './register.cypress'
+
 export * from './api/api.command'
 export * from './deprecated'
-export * from './utils.cypress'
