+++ libs/frontend/test/cypress/utils/src/register.cypress.ts
